+++ app/Http/Controllers/InstitutionController.php
@@ -67,7 +67,9 @@ class InstitutionController extends Controller
             'url_portrait'=>'mimes:jpg,png,jpeg|max:150'
             ]);
             //Elimina el documento anterior
-            unlink(public_path().'/assets/uploads/'.$institution->url_logo);
+            if(file_exists('/assets/uploads/'.$institution->url_logo)){
+                unlink(public_path().'/assets/uploads/'.$institution->url_logo);
+            }
             //Recuperando extensión de la nueva imagen
             $url_portrait = $request->file('url_portrait');
             $nombrefinal = $this->str_unico(8).'.'.$url_portrait->getClientOriginalExtension();
@@ -95,7 +97,9 @@ class InstitutionController extends Controller
     {
         $institution = Institution::find($id);
         //Eliminando el archivo
-        unlink(public_path().'/assets/uploads/'.$institution->url_logo);
+        if(file_exists('/assets/uploads/'.$institution->url_logo)){
+            unlink(public_path().'/assets/uploads/'.$institution->url_logo);
+        }
         $institution->delete();
         return redirect('panel/institution')->with('Mensaje','Cliente eliminado correctamente');
     }
