+++ app/Http/Controllers/PictureController.php
@@ -97,7 +97,9 @@ class PictureController extends Controller
     {
         $picture = Picture::find($id);
         //Eliminando el archivo
-        unlink(public_path().'/assets/uploads/'.$picture->url_picture);
+        if(file_exists('/assets/uploads/'.$picture->url_picture)){
+            unlink(public_path().'/assets/uploads/'.$picture->url_picture);
+        }        
         $picture->delete();
         return redirect('panel/picture')->with('Mensaje','Imagen eliminada correctamente');
     }
